Remove create method
