+++ Demo/TabPageViewControllerDemo/InfiniteTabPageViewController.swift
@@ -1,0 +1,35 @@
+//
+//  InfiniteTabPageViewController.swift
+//  TabPageViewController
+//
+//  Created by Tomoya Hayakawa on 2017/08/05.
+//
+//
+
+import UIKit
+import TabPageViewController
+
+class InfiniteTabPageViewController: TabPageViewController {
+    
+    override init() {
+        super.init()
+        let vc1 = UIViewController()
+        vc1.view.backgroundColor = UIColor(red: 251/255, green: 252/255, blue: 149/255, alpha: 1.0)
+        let vc2 = UIViewController()
+        vc2.view.backgroundColor = UIColor(red: 252/255, green: 150/255, blue: 149/255, alpha: 1.0)
+        let vc3 = UIViewController()
+        vc3.view.backgroundColor = UIColor(red: 149/255, green: 218/255, blue: 252/255, alpha: 1.0)
+        let vc4 = UIViewController()
+        vc4.view.backgroundColor = UIColor(red: 149/255, green: 252/255, blue: 197/255, alpha: 1.0)
+        let vc5 = UIViewController()
+        vc5.view.backgroundColor = UIColor(red: 252/255, green: 182/255, blue: 106/255, alpha: 1.0)
+        tabItems = [(vc1, "Mon."), (vc2, "Tue."), (vc3, "Wed."), (vc4, "Thu."), (vc5, "Fri.")]
+        isInfinity = true
+        option.currentColor = UIColor(red: 246/255, green: 175/255, blue: 32/255, alpha: 1.0)
+        option.tabMargin = 30.0
+    }
+
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
