+++ Demo/TabPageViewControllerDemo/ViewController.swift
@@ -11,49 +11,13 @@ import TabPageViewController
 
 class ViewController: UIViewController {
 
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
     @IBAction func LimitedButton(_ button: UIButton) {
-        let tc = TabPageViewController.create()
-        let vc1 = UIViewController()
-        vc1.view.backgroundColor = UIColor.white
-        let vc2 = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "ListViewController")
-        tc.tabItems = [(vc1, "First"), (vc2, "Second")]
-        var option = TabPageOption()
-        option.tabWidth = view.frame.width / CGFloat(tc.tabItems.count)
-        option.hidesTopViewOnSwipeType = .all
-        tc.option = option
+        let tc = LimitedTabPageViewController()
         navigationController?.pushViewController(tc, animated: true)
     }
 
     @IBAction func InfinityButton(_ button: UIButton) {
-        let tc = TabPageViewController.create()
-        let vc1 = UIViewController()
-        vc1.view.backgroundColor = UIColor(red: 251/255, green: 252/255, blue: 149/255, alpha: 1.0)
-        let vc2 = UIViewController()
-        vc2.view.backgroundColor = UIColor(red: 252/255, green: 150/255, blue: 149/255, alpha: 1.0)
-        let vc3 = UIViewController()
-        vc3.view.backgroundColor = UIColor(red: 149/255, green: 218/255, blue: 252/255, alpha: 1.0)
-        let vc4 = UIViewController()
-        vc4.view.backgroundColor = UIColor(red: 149/255, green: 252/255, blue: 197/255, alpha: 1.0)
-        let vc5 = UIViewController()
-        vc5.view.backgroundColor = UIColor(red: 252/255, green: 182/255, blue: 106/255, alpha: 1.0)
-        tc.tabItems = [(vc1, "Mon."), (vc2, "Tue."), (vc3, "Wed."), (vc4, "Thu."), (vc5, "Fri.")]
-        tc.isInfinity = true
-        let nc = UINavigationController()
-        nc.viewControllers = [tc]
-        var option = TabPageOption()
-        option.currentColor = UIColor(red: 246/255, green: 175/255, blue: 32/255, alpha: 1.0)
-        option.tabMargin = 30.0
-        tc.option = option
+        let tc = InfiniteTabPageViewController()
         navigationController?.pushViewController(tc, animated: true)
     }
 }
